requirements.txt
Pillow
opencv-python
arabic-reshaper

